Fix line intersect bug
I found this bug on Android, but probably will occur on the Java for desktop also.

With those points (in code for facility):

Polygon polygon = Polygon.Builder()
        .addVertex(new Point(399, 115))
        .addVertex(new Point(658, 115))
        .addVertex(new Point(658, 373))
        .addVertex(new Point(462, 569))
        .addVertex(new Point(203, 569))
        .addVertex(new Point(203, 311))
        .build();

Calling polygon.contains(new Point(493, 266)) returns false.
But the Point is clearly inside. 

Running the debugger I found that the tx and ty, at the isInside(Point) method, are slightly the 
same! 
The values are:
tx = 0.7093973
ty = 0.7093975

With my modification, we do not have the "almost" intercept problem - 
caused by the float precision.